+++ docs/versionhistory.rst
@@ -19,6 +19,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - **BACKWARD INCOMPATIBLE** Dropped the ``Context.loop`` attribute
 - **BACKWARD INCOMPATIBLE** Dropped the deprecated ability to use a ``Context`` as a
   synchronous context manager
+- **BACKWARD INCOMPATIBLE** Dropped the deprecated ``parent`` argument to ``Context``
 - **BACKWARD INCOMPATIBLE** Dropped the ``--unsafe`` switch for ``asphalt run`` –
   configuration files are now always parsed in unsafe mode
   (use the ``!!python/name`` YAML tag instead and replace ``:`` with ``.`` in the